init Methods changed

+++ srcs/Server.cpp
@@ -38,7 +38,7 @@ int Server::getSd() const
 void Server::setSd(int sd)
 	{ _sd = sd; }
 
-size_t Server::getServNbr() const
+int Server::getServNbr() const
 	{ return _srvrNbr; }
 
 void Server::setSrvrNbr(size_t indx)
